+++ extensions/Worker.Extensions.CosmosDB/release_notes.md
@@ -6,5 +6,6 @@
 
 ### Microsoft.Azure.Functions.Worker.Extensions.CosmosDB 4.16.0
 
-- chore: updates dependency `Microsoft.Azure.WebJobs.Extensions.CosmosDB` to version 4.16.0 (#3420)
 - feat: support `AllVersionsAndDeletes` changefeed mode (#3420)
+- build: update `Microsoft.Azure.WebJobs.Extensions.CosmosDB` to 4.16.0 (#3420)
+- build: update `Microsoft.Azure.Cosmos` to 3.60.0 (#3421)